+++ src/Umbraco.Web.Common/Security/ConfigureKestrelServerOptions.cs
@@ -15,6 +15,6 @@ public class ConfigureKestrelServerOptions : IConfigureOptions<KestrelServerOpti
 
         // convert from KB to bytes, 52428800 bytes (50 MB) is the same as in the IIS settings
         options.Limits.MaxRequestBodySize = _runtimeSettings.Value.MaxRequestLength.HasValue
-            ? _runtimeSettings.Value.MaxRequestLength.Value * 1024
+            ? (long)_runtimeSettings.Value.MaxRequestLength.Value * 1024
             : 52428800;
 }
